refactor: make jei integration common
closes #325
