+++ refinedstorage2-platform-common/build.gradle
@@ -5,6 +5,13 @@ plugins {
 
 archivesBaseName = 'refinedstorage2-platform-common'
 
+repositories {
+    maven {
+        name = 'JEI'
+        url = "https://maven.blamejared.com/"
+    }
+}
+
 minecraft {
     version(libs.versions.minecraft.get())
 }
@@ -16,6 +23,7 @@ sourceSets {
 dependencies {
     compileOnly libs.mixin
     implementation libs.apiguardian // mandatory as we import it from platform-api
+    implementation libs.jei.common.api
     implementation project(':refinedstorage2-platform-api')
     implementation project(':refinedstorage2-core-api')
     implementation project(':refinedstorage2-resource-api')
