+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/integration/jei/JeiProxy.java
@@ -1,10 +1,10 @@
-package com.refinedmods.refinedstorage2.platform.forge.integration.jei;
+package com.refinedmods.refinedstorage2.platform.common.integration.jei;
 
 import mezz.jei.api.runtime.IJeiRuntime;
 
 public class JeiProxy {
     public String getSearchFieldText() {
-        final IJeiRuntime runtime = JeiPlugin.getRuntime();
+        final IJeiRuntime runtime = RefinedStorageModPlugin.getRuntime();
         if (runtime == null) {
             return "";
         }
@@ -12,7 +12,7 @@ public class JeiProxy {
     }
 
     public void setSearchFieldText(final String text) {
-        final IJeiRuntime runtime = JeiPlugin.getRuntime();
+        final IJeiRuntime runtime = RefinedStorageModPlugin.getRuntime();
         if (runtime != null) {
             runtime.getIngredientFilter().setFilterText(text);
         }
