+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/integration/jei/RefinedStorageModPlugin.java
@@ -1,14 +1,16 @@
-package com.refinedmods.refinedstorage2.platform.fabric.integration.jei;
+package com.refinedmods.refinedstorage2.platform.common.integration.jei;
 
 import javax.annotation.Nullable;
 
 import mezz.jei.api.IModPlugin;
+import mezz.jei.api.JeiPlugin;
 import mezz.jei.api.runtime.IJeiRuntime;
 import net.minecraft.resources.ResourceLocation;
 
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createIdentifier;
 
-public class JeiPlugin implements IModPlugin {
+@JeiPlugin
+public class RefinedStorageModPlugin implements IModPlugin {
     private static final ResourceLocation ID = createIdentifier("plugin");
 
     @Nullable
@@ -21,7 +23,7 @@ public class JeiPlugin implements IModPlugin {
 
     @Override
     public void onRuntimeAvailable(final IJeiRuntime newRuntime) {
-        JeiPlugin.runtime = newRuntime;
+        RefinedStorageModPlugin.runtime = newRuntime;
     }
 
     @Nullable
