--- refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/integration/jei/JeiProxy.java
@@ -1,20 +1,0 @@
-package com.refinedmods.refinedstorage2.platform.fabric.integration.jei;
-
-import mezz.jei.api.runtime.IJeiRuntime;
-
-public class JeiProxy {
-    public String getSearchFieldText() {
-        final IJeiRuntime runtime = JeiPlugin.getRuntime();
-        if (runtime == null) {
-            return "";
-        }
-        return runtime.getIngredientFilter().getFilterText();
-    }
-
-    public void setSearchFieldText(final String text) {
-        final IJeiRuntime runtime = JeiPlugin.getRuntime();
-        if (runtime != null) {
-            runtime.getIngredientFilter().setFilterText(text);
-        }
-    }
-}
