+++ refinedstorage2-platform-fabric/src/main/resources/fabric.mod.json
@@ -26,7 +26,7 @@
       "com.refinedmods.refinedstorage2.platform.fabric.integration.ModMenuIntegration"
     ],
     "jei_mod_plugin": [
-      "com.refinedmods.refinedstorage2.platform.fabric.integration.jei.JeiPlugin"
+      "com.refinedmods.refinedstorage2.platform.common.integration.jei.RefinedStorageModPlugin"
     ]
   },
   "custom": {
