--- refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/integration/jei/JeiPlugin.java
@@ -1,32 +1,0 @@
-package com.refinedmods.refinedstorage2.platform.forge.integration.jei;
-
-import javax.annotation.Nullable;
-
-import mezz.jei.api.IModPlugin;
-import mezz.jei.api.runtime.IJeiRuntime;
-import net.minecraft.resources.ResourceLocation;
-
-import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createIdentifier;
-
-@mezz.jei.api.JeiPlugin
-public class JeiPlugin implements IModPlugin {
-    private static final ResourceLocation ID = createIdentifier("plugin");
-
-    @Nullable
-    private static IJeiRuntime runtime;
-
-    @Override
-    public ResourceLocation getPluginUid() {
-        return ID;
-    }
-
-    @Override
-    public void onRuntimeAvailable(final IJeiRuntime newRuntime) {
-        JeiPlugin.runtime = newRuntime;
-    }
-
-    @Nullable
-    public static IJeiRuntime getRuntime() {
-        return runtime;
-    }
-}
